Modularize SSO provider logic
