+++ backend/src/utils/auth/passport/index.ts
@@ -1,0 +1,4 @@
+export { initializeGoogleStrategy } from "./google";
+export { initializeGitHubStrategy } from "./github";
+export { initializeGitLabStrategy } from "./gitlab";
+export { initializeSamlStrategy } from "./saml";
