+++ backend/src/utils/setup/index.ts
@@ -11,7 +11,6 @@ import {
   backfillBots,
   backfillEncryptionMetadata,
   backfillIntegration,
-  backfillPermission,
   backfillSecretBlindIndexData,
   backfillSecretFolders,
   backfillSecretVersions,
@@ -27,7 +26,12 @@ import {
   reencryptSecretBlindIndexDataSalts
 } from "./reencryptData";
 import { getMongoURL, getNodeEnv, getRedisUrl, getSentryDSN } from "../../config";
-import { initializePassport } from "../auth";
+import {
+  initializeGitHubStrategy,
+  initializeGitLabStrategy,
+  initializeGoogleStrategy,
+  initializeSamlStrategy
+} from "../auth/passport";
 
 /**
  * Prepare Infisical upon startup. This includes tasks like:
@@ -41,6 +45,7 @@ import { initializePassport } from "../auth";
  */
 export const setup = async () => {
   if ((await getRedisUrl()) === undefined || (await getRedisUrl()) === "") {
+    // eslint-disable-next-line no-console
     console.error(
       "WARNING: Redis is not yet configured. Infisical may not function as expected without it."
     );
@@ -55,7 +60,11 @@ export const setup = async () => {
   // initializing global feature set
   await EELicenseService.initGlobalFeatureSet();
 
-  await initializePassport();
+  // initializing auth strategies
+  await initializeGoogleStrategy();
+  await initializeGitHubStrategy()
+  await initializeGitLabStrategy();
+  await initializeSamlStrategy();
 
   // re-encrypt any data previously encrypted under server hex 128-bit ENCRYPTION_KEY
   // to base64 256-bit ROOT_ENCRYPTION_KEY
